Added label choosing option type and rho test now outputs a text file for records, too

+++ rho_test.py
@@ -22,6 +22,7 @@ GAMMA = 0.0
 DATAFILE = ''
 MULTIPLIER = 0
 RHORANGESTRING = ''
+LABEL_CHOOSER = ''
 # default rho range
 RHORANGE = np.arange(0.01, .21, .02)
 # use linear or exponential plotting
@@ -33,7 +34,7 @@ def run(rows, metric_window):
 	for rho in RHORANGE:
 		# create model
 		print 'rho='+str(rho)
-		BISmodel = AdaBanditBoost(loss=LOSS, gamma=GAMMA, rho=rho)
+		BISmodel = AdaBanditBoost(loss=LOSS, gamma=GAMMA, rho=rho, label_chooser=LABEL_CHOOSER)
 		BISmodel.M = 100
 		BISmodel.initialize_dataset(filename, class_index,
 									probe_instances=N*MULTIPLIER)
@@ -51,13 +52,22 @@ def run(rows, metric_window):
 def plotRun(rhos, accuracies):
 	specs = '_loss='+str(LOSS)+'_num_wls='+str(NUM_WLS)+\
 				'_gamma='+str(GAMMA)+'\n_DATAFILE='+DATAFILE+\
-				'_DATAMULTIPLIER='+str(MULTIPLIER)+'_arrayspec='+RHORANGESTRING
+				'_DATAMULTIPLIER='+str(MULTIPLIER)+\
+				'_labelchooser='+LABEL_CHOOSER+'_arrayspec='+RHORANGESTRING
 	if LIN_OR_EXP is 'exp':
 		# exponential plotting
 		rhos = [math.log10(r) for r in rhos]
-	print('rhos:', rhos)
+	print 'rhos:', rhos
 
+	mean_coordinates = zip(*sns.tsplot(accuracies, time=rhos).get_lines()[0].get_data())
+	# calculate standard deviation to label the plots with
+	stdevs = [np.std(v) for v in zip(*accuracies)]
+	notations = [(m[0], m[1], s) for m, s in zip(mean_coordinates, stdevs)]
+	text = '(log10 rho, mean, stdev)\n' + str(notations)
+
+	# just testing something here
 	sns.tsplot(accuracies, time=rhos)
+
 	if LIN_OR_EXP is 'exp':
 		plt.xlabel('log_10(rhos)')
 	else:
@@ -70,11 +80,20 @@ def plotRun(rhos, accuracies):
 	filename_identifier = specs.replace('\n', '')
 	savefile = os.path.join(RESULTSDIR, OUTFILENAMEBASE+\
 				filename_identifier+OUTFILETYPE)
+	savetext = os.path.join(RESULTSDIR, OUTFILENAMEBASE+\
+				filename_identifier+'.txt')
 	print 'saving graph to', savefile
+	print 'savign text to', savetext
 	if os.path.exists(savefile):
 		print 'error:', savefile, 'already exists'
 		exit(1)
+	if os.path.exists(savetext):
+		print 'error:', savetext, 'already exists'
+		exit(1)
 	plt.savefig(savefile)
+	with open(savetext, 'w') as f:
+		f.write(text)
+
 
 OUTFILENAMEBASE = 'rho_test_out'
 OUTFILETYPE = config.OUTFILETYPE
@@ -82,7 +101,7 @@ DATADIR = config.DATADIR
 RESULTSDIR = config.RESULTSDIR
 '''
 to run, call:
-python2 rho_test.py --loss=zero_one --num_wls=20 --gamma=0.1 --datafile=balance-scale.csv --multiplier=1 --array_spec=0.1_1_.4
+python2 rho_test.py --loss=logistic --num_wls=20 --gamma=0.1 --datafile=balance-scale.csv --multiplier=1 --array_spec=0.1_0.6_.4
 '''
 
 if __name__ == '__main__':
@@ -92,6 +111,7 @@ if __name__ == '__main__':
 	parser.add_argument('--gamma', type=float, required=True)
 	parser.add_argument('--datafile', type=str, required=True)
 	parser.add_argument('--multiplier', type=int, required=True)
+	parser.add_argument('--label_chooser', type=str, required=True)
 	'''
 	the way to use this argument is like 0.1,0.2,0.3 (list of values)
 	or 0.1_0.2_20 (use numpy.arange)
@@ -105,6 +125,7 @@ if __name__ == '__main__':
 	DATAFILE = args.datafile
 	MULTIPLIER = args.multiplier
 	RHORANGESTRING = args.array_spec
+	LABEL_CHOOSER = args.label_chooser
 	if '_' in args.array_spec:
 		LIN_OR_EXP = 'lin'
 		arange_args = [float(arg) for arg in args.array_spec.split('_')]
@@ -125,5 +146,5 @@ if __name__ == '__main__':
 
 	print 'running ...'
 	rho_runs = [run(utils.shuffle(rows, seed=random.randint(1, 2000000)),
-					test_N)[-1] for _ in range(5)]
+					test_N)[-1] for _ in range(20)]
 	plotRun(RHORANGE, rho_runs)
